refactor: move tests to a new file and clean up original file
